Rename fused_marlin_moe kernels by precision scheme and split benchmarks
