+++ src/flag_gems/fused/fused_marlin_moe.py
@@ -88,13 +88,13 @@ def _is_hopper() -> bool:
     return torch.cuda.is_available() and torch.cuda.get_device_capability()[0] >= 9
 
 
-class _W4A16KernelPolicy(NamedTuple):
+class _W4A16Int4KernelPolicy(NamedTuple):
     block_m: int
     use_fused_gemm1_silu: bool
     move_router_weight_before_gemm2: bool
 
 
-class _MXFP4KernelPolicy(NamedTuple):
+class _W4A16Mxfp4KernelPolicy(NamedTuple):
     block_m: int
     use_fused_gemm1_silu: bool
     align_mode: "_MXFP4AlignMode"
@@ -170,14 +170,14 @@ def _select_block_m(
     return 64
 
 
-def _select_w4a16_kernel_policy(
+def _select_w4a16_int4_kernel_policy(
     device: torch.device,
     M: int,
     E: int,
     top_k: int,
     swap_ab: bool,
     apply_router_weight_on_input: bool,
-) -> _W4A16KernelPolicy:
+) -> _W4A16Int4KernelPolicy:
     device_info = _get_device_info(device)
 
     # Base tiling policy. Full Hopper uses a smaller cutoff because it has
@@ -210,20 +210,20 @@ def _select_w4a16_kernel_policy(
         use_fused_gemm1_silu and not apply_router_weight_on_input and M >= 512
     )
 
-    return _W4A16KernelPolicy(
+    return _W4A16Int4KernelPolicy(
         block_m=block_m,
         use_fused_gemm1_silu=use_fused_gemm1_silu,
         move_router_weight_before_gemm2=move_router_weight_before_gemm2,
     )
 
 
-def _select_mxfp4_kernel_policy(
+def _select_w4a16_mxfp4_kernel_policy(
     device: torch.device,
     M: int,
     E: int,
     top_k: int,
     swap_ab: bool,
-) -> _MXFP4KernelPolicy:
+) -> _W4A16Mxfp4KernelPolicy:
     device_info = _get_device_info(device)
     is_reduced_hopper = (
         device_info.is_hopper and not device_info.has_full_hopper_sm_count
@@ -248,7 +248,7 @@ def _select_mxfp4_kernel_policy(
     else:
         align_mode = _MXFP4AlignMode.default
 
-    return _MXFP4KernelPolicy(
+    return _W4A16Mxfp4KernelPolicy(
         block_m=block_m,
         use_fused_gemm1_silu=is_reduced_hopper and M > 1,
         align_mode=align_mode,
@@ -345,7 +345,7 @@ def _cached_pack_scale(s: torch.Tensor, cached: bool) -> torch.Tensor:
     return packed
 
 
-def w4a16_pack(
+def w4a16_int4_pack(
     w1: torch.Tensor,
     w2: torch.Tensor,
     w1_scale: Optional[torch.Tensor] = None,
@@ -437,7 +437,7 @@ def _cached_pack_scale_e8m0_fold(s, compute_dtype, cached: bool) -> torch.Tensor
     return packed
 
 
-def mxfp4_pack(
+def w4a16_mxfp4_pack(
     w1,
     w2,
     w1_scale,
@@ -852,7 +852,7 @@ def _write_zeros_to_output(
     ],
 )
 @triton.jit
-def _w4a16_moe_gemm_kernel(
+def _w4a16_int4_moe_gemm_kernel(
     a_ptr,
     b_ptr,
     c_ptr,
@@ -1022,7 +1022,7 @@ def _w4a16_moe_gemm_kernel(
     ],
 )
 @triton.jit
-def _w4a16_moe_gemm_silu_kernel(
+def _w4a16_int4_moe_gemm_silu_kernel(
     a_ptr,
     b_ptr,
     c_ptr,
@@ -1193,7 +1193,7 @@ def _w4a16_moe_gemm_silu_kernel(
     tl.store(c_ptrs, accumulator, mask=c_mask)
 
 
-def _invoke_w4a16_moe_gemm(
+def _invoke_w4a16_int4_moe_gemm(
     A: torch.Tensor,  # (M, K) for GEMM1, (M*top_k, K) for GEMM2
     B: torch.Tensor,  # (E, K//8, N) int32
     C: torch.Tensor,  # (M, top_k, N) or (M*top_k, N) view
@@ -1229,7 +1229,7 @@ def _invoke_w4a16_moe_gemm(
         triton.cdiv(EM, META["BLOCK_SIZE_M"]) * triton.cdiv(N, META["BLOCK_SIZE_N"]),
     )
 
-    _w4a16_moe_gemm_kernel[grid](
+    _w4a16_int4_moe_gemm_kernel[grid](
         A,
         B,
         C,
@@ -1262,7 +1262,7 @@ def _invoke_w4a16_moe_gemm(
     )
 
 
-def _invoke_w4a16_moe_gemm_silu(
+def _invoke_w4a16_int4_moe_gemm_silu(
     A: torch.Tensor,
     B: torch.Tensor,
     C: torch.Tensor,
@@ -1291,7 +1291,7 @@ def _invoke_w4a16_moe_gemm_silu(
         triton.cdiv(EM, META["BLOCK_SIZE_M"]) * triton.cdiv(N, META["BLOCK_SIZE_N"]),
     )
 
-    _w4a16_moe_gemm_silu_kernel[grid](
+    _w4a16_int4_moe_gemm_silu_kernel[grid](
         A,
         B,
         C,
@@ -1329,7 +1329,7 @@ def _invoke_w4a16_moe_gemm_silu(
     )
 
 
-def fused_moe_w4a16_gptq(
+def fused_marlin_moe_w4a16_int4(
     hidden_states: torch.Tensor,
     w1: torch.Tensor,
     w2: torch.Tensor,
@@ -1373,7 +1373,7 @@ def fused_moe_w4a16_gptq(
     else:
         compute_type = tl.bfloat16
 
-    w1_packed, w2_packed, w1_scale_packed, w2_scale_packed = w4a16_pack(
+    w1_packed, w2_packed, w1_scale_packed, w2_scale_packed = w4a16_int4_pack(
         w1,
         w2,
         w1_scale,
@@ -1382,7 +1382,7 @@ def fused_moe_w4a16_gptq(
         cached=True,
     )
 
-    policy = _select_w4a16_kernel_policy(
+    policy = _select_w4a16_int4_kernel_policy(
         hidden_states.device,
         M,
         E,
@@ -1424,7 +1424,7 @@ def fused_moe_w4a16_gptq(
     )
 
     if use_fused_gemm1_silu:
-        _invoke_w4a16_moe_gemm_silu(
+        _invoke_w4a16_int4_moe_gemm_silu(
             A=hidden_states,
             B=w1_packed,
             C=intermediate_cache2,
@@ -1447,7 +1447,7 @@ def fused_moe_w4a16_gptq(
         )
     else:
         assert intermediate_cache1 is not None
-        _invoke_w4a16_moe_gemm(
+        _invoke_w4a16_int4_moe_gemm(
             A=hidden_states,
             B=w1_packed,
             C=intermediate_cache1,
@@ -1473,7 +1473,7 @@ def fused_moe_w4a16_gptq(
     else:
         out_hidden_states = torch.empty_like(hidden_states)
 
-    _invoke_w4a16_moe_gemm(
+    _invoke_w4a16_int4_moe_gemm(
         A=intermediate_cache2,
         B=w2_packed,
         C=intermediate_cache3,
@@ -1498,14 +1498,14 @@ def fused_moe_w4a16_gptq(
 
 @libentry()
 @libtuner(
-    configs=runtime.get_tuned_config("fused_marlin_moe_mxfp4"),
+    configs=runtime.get_tuned_config("fused_marlin_moe_w4a16_mxfp4"),
     key=["N", "K", "EM_BUCKET", "BLOCK_SIZE_M", "SWAP_AB"],
     strategy=["align32", "align32", "align32", "align32", "default"],
-    flagtune_op_name="fused_marlin_moe_mxfp4",
-    flagtune_expand_op_name="fused_marlin_moe_mxfp4",
+    flagtune_op_name="fused_marlin_moe_w4a16_mxfp4",
+    flagtune_expand_op_name="fused_marlin_moe_w4a16_mxfp4",
 )
 @triton.jit
-def _mxfp4_moe_gemm_kernel(
+def _w4a16_mxfp4_moe_gemm_kernel(
     a_ptr,
     b_ptr,
     c_ptr,
@@ -1657,14 +1657,14 @@ def _mxfp4_moe_gemm_kernel(
 
 @libentry()
 @libtuner(
-    configs=runtime.get_tuned_config("fused_marlin_moe_mxfp4_gemm_silu"),
+    configs=runtime.get_tuned_config("fused_marlin_moe_w4a16_mxfp4_gemm_silu"),
     key=["N", "K", "BLOCK_SIZE_M", "SWAP_AB"],
     strategy=["align32", "align32", "align32", "default"],
-    flagtune_op_name="fused_marlin_moe_mxfp4_gemm_silu",
-    flagtune_expand_op_name="fused_marlin_moe_mxfp4_gemm_silu",
+    flagtune_op_name="fused_marlin_moe_w4a16_mxfp4_gemm_silu",
+    flagtune_expand_op_name="fused_marlin_moe_w4a16_mxfp4_gemm_silu",
 )
 @triton.jit
-def _mxfp4_moe_gemm_silu_kernel(
+def _w4a16_mxfp4_moe_gemm_silu_kernel(
     a_ptr,
     b_ptr,
     c_ptr,
@@ -1863,7 +1863,7 @@ def _mxfp4_moe_gemm_silu_kernel(
     tl.store(c_ptrs, accumulator, mask=c_mask)
 
 
-def _invoke_mxfp4_moe_gemm(
+def _invoke_w4a16_mxfp4_moe_gemm(
     A,
     B,
     C,
@@ -1901,7 +1901,7 @@ def _invoke_mxfp4_moe_gemm(
         triton.cdiv(EM, META["BLOCK_SIZE_M"]) * triton.cdiv(N, META["BLOCK_SIZE_N"]),
     )
 
-    _mxfp4_moe_gemm_kernel[grid](
+    _w4a16_mxfp4_moe_gemm_kernel[grid](
         A,
         B,
         C,
@@ -1936,7 +1936,7 @@ def _invoke_mxfp4_moe_gemm(
     )
 
 
-def _invoke_mxfp4_moe_gemm_silu(
+def _invoke_w4a16_mxfp4_moe_gemm_silu(
     A,
     B,
     C,
@@ -1966,7 +1966,7 @@ def _invoke_mxfp4_moe_gemm_silu(
         triton.cdiv(EM, META["BLOCK_SIZE_M"]) * triton.cdiv(N, META["BLOCK_SIZE_N"]),
     )
 
-    _mxfp4_moe_gemm_silu_kernel[grid](
+    _w4a16_mxfp4_moe_gemm_silu_kernel[grid](
         A,
         B,
         C,
@@ -2000,7 +2000,7 @@ def _invoke_mxfp4_moe_gemm_silu(
     )
 
 
-def fused_moe_mxfp4(
+def fused_marlin_moe_w4a16_mxfp4(
     hidden_states: torch.Tensor,
     w1: torch.Tensor,
     w2: torch.Tensor,
@@ -2049,7 +2049,7 @@ def fused_moe_mxfp4(
         and _e8m0_fold_safe(w2_scale, hidden_states.dtype)
     )
 
-    w1_packed, w2_packed, w1_scale_packed, w2_scale_packed = mxfp4_pack(
+    w1_packed, w2_packed, w1_scale_packed, w2_scale_packed = w4a16_mxfp4_pack(
         w1,
         w2,
         w1_scale,
@@ -2060,7 +2060,7 @@ def fused_moe_mxfp4(
         fold_scale=fold_scale,
     )
 
-    policy = _select_mxfp4_kernel_policy(
+    policy = _select_w4a16_mxfp4_kernel_policy(
         hidden_states.device,
         M,
         E,
@@ -2095,7 +2095,7 @@ def fused_moe_mxfp4(
     )
 
     if use_fused_gemm1_silu:
-        _invoke_mxfp4_moe_gemm_silu(
+        _invoke_w4a16_mxfp4_moe_gemm_silu(
             A=hidden_states,
             B=w1_packed,
             C=intermediate_cache2,
@@ -2115,7 +2115,7 @@ def fused_moe_mxfp4(
         )
     else:
         assert intermediate_cache1 is not None
-        _invoke_mxfp4_moe_gemm(
+        _invoke_w4a16_mxfp4_moe_gemm(
             A=hidden_states,
             B=w1_packed,
             C=intermediate_cache1,
@@ -2138,7 +2138,7 @@ def fused_moe_mxfp4(
         up = intermediate_cache1[:, intermediate_size:]
         silu_and_mul_out(gate, up, intermediate_cache2)
 
-    _invoke_mxfp4_moe_gemm(
+    _invoke_w4a16_mxfp4_moe_gemm(
         A=intermediate_cache2,
         B=w2_packed,
         C=intermediate_cache3,
@@ -2504,7 +2504,7 @@ def fused_marlin_moe(
         and w1_scale.dtype == hidden_states.dtype
         and w2_scale.dtype == hidden_states.dtype
     ):
-        result = fused_moe_w4a16_gptq(
+        result = fused_marlin_moe_w4a16_int4(
             hidden_states=hidden_states,
             w1=w1,
             w2=w2,
@@ -2540,7 +2540,7 @@ def fused_marlin_moe(
                 "MXFP4 fast path requires Hopper, bf16/fp16 activations, uint8 "
                 "packed weights, no bias/zeros/expert_map."
             )
-        result = fused_moe_mxfp4(
+        result = fused_marlin_moe_w4a16_mxfp4(
             hidden_states=hidden_states,
             w1=w1,
             w2=w2,
