+++ src/flag_gems/runtime/common.py
@@ -48,14 +48,14 @@ DEFAULT_STRATEGIES = {
     "bmm": ["align32", "align32", "align32", "align32", "align32"],
     "bmm_sqmma": ["align32", "align32", "align32"],
     "compute_global_topk_indices_and_lens": ["align32", "align32"],
-    "fused_marlin_moe_mxfp4": [
+    "fused_marlin_moe_w4a16_mxfp4": [
         "align32",
         "align32",
         "align32",
         "align32",
         "default",
     ],
-    "fused_marlin_moe_mxfp4_gemm_silu": [
+    "fused_marlin_moe_w4a16_mxfp4_gemm_silu": [
         "align32",
         "align32",
         "align32",
@@ -127,14 +127,14 @@ OP_KEY_ORDERS = {
     "bmm_sqmma": ["M", "N", "K"],
     "baddbmm": ["M", "N", "K"],
     "compute_global_topk_indices_and_lens": ["topk", "num_tokens"],
-    "fused_marlin_moe_mxfp4": [
+    "fused_marlin_moe_w4a16_mxfp4": [
         "N",
         "K",
         "EM_BUCKET",
         "BLOCK_SIZE_M",
         "SWAP_AB",
     ],
-    "fused_marlin_moe_mxfp4_gemm_silu": [
+    "fused_marlin_moe_w4a16_mxfp4_gemm_silu": [
         "N",
         "K",
         "BLOCK_SIZE_M",
